# backend/docs/swagger.yaml
basePath: /
definitions:
...
  dtos.DepositDto:
    properties:
      account_id:
        example: acc_123
        type: string
      amount:
        example: 2500
        type: number
      description:
        example: Salary deposit
        type: string
    required:
    - account_id
    - amount
    type: object
  dtos.DepositResponse:
    properties:
      message:
        type: string
      payload:
        $ref: '#/definitions/dtos.TransactionResponse'
      status:
        type: integer
      success:
        type: boolean
    type: object
  dtos.EntryResponse:
    properties:
      account_id:
        type: string
      created_at:
        type: string
      credit:
        type: string
      debit:
        type: string
      description:
        type: string
      id:
        type: string
    type: object
...
  dtos.GetTransactionResponse:
    properties:
      message:
        type: string
      payload:
        $ref: '#/definitions/dtos.TransactionResponse'
      status:
        type: integer
      success:
        type: boolean
    type: object
...
  dtos.ListEntriesResponse:
    properties:
      message:
        type: string
      payload:
        items:
          $ref: '#/definitions/dtos.EntryResponse'
        type: array
      status:
        type: integer
      success:
        type: boolean
    type: object
  dtos.ListTransactionsResponse:
    properties:
      message:
        type: string
      payload:
        items:
          $ref: '#/definitions/dtos.TransactionResponse'
        type: array
      status:
        type: integer
      success:
        type: boolean
    type: object
...
  dtos.ReconcileAccountResponse:
    properties:
      message:
        type: string
      payload:
        $ref: '#/definitions/dtos.ReconcileResponse'
      status:
        type: integer
      success:
        type: boolean
    type: object
  dtos.ReconcileDto:
    properties:
      account_id:
        type: string
      expected_balance:
        minimum: 0
        type: number
    required:
    - account_id
    - expected_balance
    type: object
  dtos.ReconcileResponse:
    properties:
      account_id:
        type: string
      calculated_balance:
        type: string
      discrepancy:
        type: string
      expected_balance:
        type: string
      reconciled:
        type: boolean
    type: object
...
  dtos.StructuredResponse:
    description: Standard response format containing success status, HTTP status code,
      message, and optional payload
    properties:
      message:
        description: |-
          Human-readable message
          @example Operation completed successfully
        example: Operation completed successfully
        type: string
      payload:
        description: |-
          Optional payload containing response data
          @example {"id":1,"name":"Example Item"}
      status:
        description: |-
          HTTP status code
          @example 200
        example: 200
        type: integer
      success:
        description: |-
          Whether the operation was successful
          @example true
        example: true
        type: boolean
    type: object
  dtos.TransactionResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      entries:
        items:
          $ref: '#/definitions/dtos.EntryResponse'
        type: array
      id:
        type: string
      operation_type:
        type: string
    type: object
  dtos.TransferDto:
    properties:
      amount:
        example: 1500
        type: number
      description:
        example: rent payment
        type: string
      from_account_id:
        example: acc_123
        type: string
      to_account_id:
        example: acc_456
        type: string
    required:
    - amount
    - from_account_id
    - to_account_id
    type: object
  dtos.TransferResponse:
    properties:
      message:
        type: string
      payload:
        $ref: '#/definitions/dtos.TransactionResponse'
      status:
        type: integer
      success:
        type: boolean
    type: object
  dtos.WithdrawDto:
    properties:
      account_id:
        example: acc_123
        type: string
      amount:
        example: 500
        type: number
      description:
        example: ATM withdrawal
        type: string
    required:
    - account_id
    - amount
    type: object
  dtos.WithdrawResponse:
    properties:
      message:
        type: string
      payload:
        $ref: '#/definitions/dtos.TransactionResponse'
      status:
        type: integer
      success:
        type: boolean
    type: object
host: localhost:8000
info:
  contact: {}
  description: Double-entry accounting ledger
  title: Double-Entry Bank Ledger API
  version: "1.0"
paths:
...
  /accounts/{account_id}/entries:
    get:
      consumes:
      - application/json
      description: Retrieves ledger entries for a specific account with pagination
        support.
      parameters:
      - description: Account ID
        in: path
        name: account_id
        required: true
        type: string
      - description: Number of records to return (1-100)
        in: query
        name: limit
        required: true
        type: integer
      - description: Pagination offset (start index)
        in: query
        name: offset
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.ListEntriesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
      security:
      - BearerAuth: []
      summary: List account entries
      tags:
      - entries
  /accounts/{account_id}/transactions:
    get:
      consumes:
      - application/json
      description: Retrieves all transactions for a specific account owned by the
        authenticated user.
      parameters:
      - description: Account ID
        in: path
        name: account_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.ListTransactionsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
      security:
      - BearerAuth: []
      summary: List account transactions
      tags:
      - transactions
...
  /accounts/reconcile:
    post:
      consumes:
      - application/json
      description: Compares expected balance with calculated ledger balance to verify
        account integrity.
      parameters:
      - description: Reconciliation payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dtos.ReconcileDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.ReconcileAccountResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
      security:
      - BearerAuth: []
      summary: Reconcile account balance
      tags:
      - accounts
...
  /transactions/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves a single transaction by ID for the authenticated user.
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.GetTransactionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
      security:
      - BearerAuth: []
      summary: Get a transaction
      tags:
      - transactions
  /transactions/deposit:
    post:
      consumes:
      - application/json
      description: Deposits money into a user account and creates a transaction record.
      parameters:
      - description: Deposit payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dtos.DepositDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dtos.DepositResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
      security:
      - BearerAuth: []
      summary: Deposit funds
      tags:
      - transactions
  /transactions/transfer:
    post:
      consumes:
      - application/json
      description: Transfers money from one account to another and creates a transaction
        record.
      parameters:
      - description: Transfer payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dtos.TransferDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dtos.TransferResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
      security:
      - BearerAuth: []
      summary: Transfer funds between accounts
      tags:
      - transactions
  /transactions/withdraw:
    post:
      consumes:
      - application/json
      description: Withdraws money from a user account and creates a transaction record.
      parameters:
      - description: Withdrawal payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dtos.WithdrawDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dtos.WithdrawResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
      security:
      - BearerAuth: []
      summary: Withdraw funds
      tags:
      - transactions
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"